Allow custom semantic-ui builds
Adds two configuration options. One for the path to the `dist` folder of
Semantic. The other option to specify the theme used for images and
fonts.

Note that this does not add support for importing files outside of the
build tree (e.g. `bower_components`). One can use `bower link` for local
projects.